refactor(threads): useThreadEventHandlers 拆诊断运行时/生命周期/延迟对账
- 2747→908 行：diagnosticsRuntime 1110 + lifecycleHandlers 684 + deferredReconciliation 317
- 9 个共享 timer refs 仅在工厂 hook 创建，聚合返回注入；55 回调逐字、35-key deps diff 为空
- useCallback 54 个一一对应；循环依赖按 section 归属解开
- baseline 重生成 + 新文件登记豁免
验证：tsc 零 error；useThreadEventHandlers.test 68 passed；gate 绿
（子代理 JudicialPuma 执行，主代理复核验证）
