+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T18:58:57.017Z",
+  "generatedAt": "2026-08-30T19:00:31.216Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
