+++ docs/architecture/large-file-new-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T18:58:57.924Z",
+  "generatedAt": "2026-08-30T19:00:32.077Z",
   "scope": "new-file",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
